+++ server/module-shared/src/main/java/project/BaekjoonStatus/shared/domain/user/repository/UserRepositoryImpl.java
@@ -1,0 +1,37 @@
+package project.BaekjoonStatus.shared.domain.user.repository;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+import project.BaekjoonStatus.shared.domain.user.entity.User;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserRepositoryImpl implements UserRepository {
+    private final UserJpaRepository userJpaRepository;
+
+    @Override
+    public boolean existByUsername(String username) {
+        return userJpaRepository.existsByUsername(username);
+    }
+
+    @Override
+    @Transactional
+    public User save(User user) {
+        return userJpaRepository.save(user);
+    }
+
+    @Override
+    public Optional<User> findById(String userId) {
+        return userJpaRepository.findById(UUID.fromString(userId));
+    }
+
+    @Override
+    public Optional<User> findByUsername(String username) {
+        return userJpaRepository.findByUsername(username);
+    }
+}
